+++ bin/transform.py
@@ -14,6 +14,8 @@ SAMPLE_HEADER = "sample"
 SAMPLE_NAME_HEADER = "sample_name"
 AGE_HEADER = "age"
 EARLIEST_HEADER = "earliest_date"
+POPULATE_HEADER = "populated"
+
 VALID_HEADER_EXTENSION = "_valid"
 ERROR_HEADER_EXTENSION = "_error"
 
@@ -25,6 +27,7 @@ DATE_2_INDEX = 3
 LOCK = "lock"
 AGE = "age"
 EARLIEST = "earliest"
+POPULATE = "populate"
 
 # Output Files:
 RESULTS_PATH = "results.csv"
@@ -35,9 +38,18 @@ DATE_FORMAT = "%Y-%m-%d" # YYYY-MM-DD
 AGE_THRESHOLD = 2 # Ages less than this will include a decimal component.
 DAYS_IN_YEAR = 365.0
 COLUMN_WISE = 0 # i.e. axis=0 // axis="columns"
+POPULATE_VALUE = "NA"
 
 def remove_empty_columns(metadata):
     metadata.dropna(axis="columns", how="all", inplace=True)
+
+def populate(metadata, populate_header, populate_value):
+    metadata_readable = metadata.copy(deep=True)
+    metadata_readable[[populate_header]] = populate_value
+
+    metadata_irida = metadata_readable[[SAMPLE_HEADER, populate_header]].copy(deep=True)
+
+    return metadata_readable, metadata_irida
 
 def find_earliest_date(row):
     earliest = ""
@@ -178,12 +190,16 @@ def main():
 
     parser.add_argument("input", type=pathlib.Path,
                         help="The CSV-formatted input file to transform.")
-    parser.add_argument("transformation", choices=[LOCK, AGE, EARLIEST],
+    parser.add_argument("transformation", choices=[LOCK, AGE, EARLIEST, POPULATE],
                         help="The type of transformation to perform.")
     parser.add_argument("--age_header", default=AGE_HEADER, required=False,
                         help="The output column header for the calculated age.")
     parser.add_argument("--earliest_header", default=EARLIEST_HEADER, required=False,
                         help="The output column header for the earliest date.")
+    parser.add_argument("--populate_header", default=POPULATE_HEADER, required=False,
+                        help="The output column header for the populate transformation.")
+    parser.add_argument("--populate_value", default=POPULATE_VALUE, required=False,
+                        help="The value to populate the specified column with for the populate transformation.")
 
     args = parser.parse_args()
     metadata = pandas.read_csv(args.input)
@@ -204,6 +220,13 @@ def main():
 
     elif (args.transformation == EARLIEST):
         metadata_readable, metadata_irida = earliest(metadata, args.earliest_header)
+
+        remove_empty_columns(metadata_irida)
+        metadata_readable.to_csv(RESULTS_PATH, index=False)
+        metadata_irida.to_csv(TRANSFORMATION_PATH, index=False)
+
+    elif (args.transformation == POPULATE):
+        metadata_readable, metadata_irida = populate(metadata, args.populate_header, args.populate_value)
 
         remove_empty_columns(metadata_irida)
         metadata_readable.to_csv(RESULTS_PATH, index=False)