+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### `Added`
 
 - `earliest` transformation: finds earliest date among metadata
+- `populate` transformation: populates an output column with a specific value
 
 ## [1.0.0] - 2025/03/07